+++ NEWS.md
@@ -76,7 +76,7 @@ Use it with `Basic.DSSCompatFlags()`.
 
 - Add Enum support for all functions
 - Minor fixes in Error module
-- Add OPENDSSDIRECT_REPL environment variable to disable REPL mode
+- Add `OPENDSSDIRECT_REPL` environment variable to disable REPL mode
 - Add download examples function
 
 ### OpenDSSDirect v0.5.0 Release Notes